Move generator and charge classifier/creator into helpers/
The root level of models/ now only contains "plain" models. Everything else including generator.py will go under models/helpers/.
